# specs/001-employee-leave-management/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Employee Leave Management API
  version: 0.1.0
  description: REST contract for the Employee Leave Management MVP. See authorization.md for row-level scope rules.
servers:
  - url: /api
security:
  - sessionCookie: []
...
paths:
...
  /employee/leave-requests/calculate:
    post:
      tags: [Employee]
      operationId: calculateOwnLeaveRequest
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LeaveRequestInput' }
      responses:
        '200':
          description: Authoritative preview; does not reserve balance
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LeaveCalculation' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/Unprocessable' }
  /employee/leave-requests:
    get:
      tags: [Employee]
      operationId: listOwnLeaveRequests
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/Status'
      responses:
        '200':
          description: Paginated own request history
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LeaveRequestPage' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [Employee]
      operationId: submitOwnLeaveRequest
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LeaveRequestInput' }
      responses:
        '201':
          description: Pending request created and balance reserved where required
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LeaveRequestDetail' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/Unprocessable' }
  /employee/leave-requests/{requestId}:
    get:
      tags: [Employee]
      operationId: getOwnLeaveRequest
      parameters:
        - $ref: '#/components/parameters/RequestId'
      responses:
        '200':
          description: Own request detail and status history
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LeaveRequestDetail' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /manager/leave-requests/{requestId}:
    get:
      tags: [Manager]
      operationId: getDirectReportLeaveRequest
      parameters:
        - $ref: '#/components/parameters/RequestId'
      responses:
        '200':
          description: Direct-report request with relevant balance
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ManagerLeaveRequestDetail' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /admin/employees/{employeeId}:
    patch:
      tags: [Administration]
      operationId: updateEmployee
      parameters:
        - $ref: '#/components/parameters/EmployeeId'
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EmployeeUpdateCommand' }
      responses:
        '200':
          description: Employee/reporting relationship updated or deactivated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Employee' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: JSESSIONID
  parameters:
    CsrfHeader:
      name: X-XSRF-TOKEN
      in: header
      required: true
      schema: { type: string }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema: { type: string, minLength: 8, maxLength: 100 }
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 0, default: 0 }
    Size:
      name: size
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    Status:
      name: status
      in: query
      schema: { $ref: '#/components/schemas/LeaveStatus' }
...
    RequestId:
      name: requestId
      in: path
      required: true
      schema: { type: string, format: uuid }
    EmployeeId:
      name: employeeId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Unauthorized:
      description: Authentication required or login failed
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Forbidden:
      description: Role, ownership, or reporting scope denied
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: Resource absent or hidden from current scope
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Conflict:
      description: Overlap, balance, stale state, policy, or transition conflict
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Unprocessable:
      description: Business rule prevents the command
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
...
    Role:
      type: string
      enum: [EMPLOYEE, MANAGER, ADMINISTRATOR]
    Roles:
      type: array
      minItems: 1
      uniqueItems: true
      items: { $ref: '#/components/schemas/Role' }
    LeaveStatus:
      type: string
      enum: [PENDING, APPROVED, REJECTED, CANCELLED]
    DurationMode:
      type: string
      enum: [FULL_DAY, HALF_DAY_AM, HALF_DAY_PM]
    LeaveRequestInput:
      type: object
      additionalProperties: false
      required: [leaveTypeId, startDate, endDate, durationMode, reason]
      properties:
        leaveTypeId: { type: string, format: uuid }
        startDate: { type: string, format: date }
        endDate: { type: string, format: date }
        durationMode: { $ref: '#/components/schemas/DurationMode' }
        reason: { type: string, minLength: 1, maxLength: 1000 }
    LeaveCalculation:
      type: object
      required: [chargeableDays, chargeableDates, tracksBalance, canSubmit]
      properties:
        chargeableDays: { type: number, multipleOf: 0.5, minimum: 0.5 }
        chargeableDates:
          type: array
          items: { type: string, format: date }
        excludedDates:
          type: array
          items: { $ref: '#/components/schemas/ExcludedDate' }
        tracksBalance: { type: boolean, description: When true, submission must validate and reserve sufficient available balance }
        availableDays: { type: number, multipleOf: 0.5 }
        canSubmit: { type: boolean }
        messages:
          type: array
          items: { type: string }
    ExcludedDate:
      type: object
      required: [date, reason]
      properties:
        date: { type: string, format: date }
        reason: { type: string, enum: [WEEKLY_OFF, HOLIDAY] }
...
    LeaveBalance:
      type: object
      required: [id, leaveTypeId, leaveTypeName, periodStart, periodEnd, entitledDays, reservedDays, consumedDays, availableDays, version]
      properties:
        id: { type: string, format: uuid }
        leaveTypeId: { type: string, format: uuid }
        leaveTypeName: { type: string }
        periodStart: { type: string, format: date }
        periodEnd: { type: string, format: date }
        entitledDays: { type: number, multipleOf: 0.5 }
        reservedDays: { type: number, multipleOf: 0.5 }
        consumedDays: { type: number, multipleOf: 0.5 }
        availableDays: { type: number, multipleOf: 0.5 }
        version: { type: integer, format: int64 }
    LeaveRequestSummary:
      type: object
      required: [id, employeeId, employeeName, leaveTypeId, leaveTypeName, startDate, endDate, durationMode, chargeableDays, status, submittedAt, version]
      properties:
        id: { type: string, format: uuid }
        employeeId: { type: string, format: uuid }
        employeeName: { type: string }
        leaveTypeId: { type: string, format: uuid }
        leaveTypeName: { type: string }
        startDate: { type: string, format: date }
        endDate: { type: string, format: date }
        durationMode: { $ref: '#/components/schemas/DurationMode' }
        chargeableDays: { type: number, multipleOf: 0.5 }
        status: { $ref: '#/components/schemas/LeaveStatus' }
        submittedAt: { type: string, format: date-time }
        version: { type: integer, format: int64 }
    LeaveRequestDetail:
      allOf:
        - $ref: '#/components/schemas/LeaveRequestSummary'
        - type: object
          required: [reason, statusHistory, canCancel]
          properties:
            reason: { type: string }
            decisionComment: { type: [string, 'null'] }
            canCancel: { type: boolean }
            cancellationBlockedReason: { type: [string, 'null'] }
            statusHistory:
              type: array
              items: { $ref: '#/components/schemas/StatusHistoryEntry' }
    ManagerLeaveRequestDetail:
      allOf:
        - $ref: '#/components/schemas/LeaveRequestDetail'
        - type: object
          properties:
            relevantBalance: { $ref: '#/components/schemas/LeaveBalance' }
    StatusHistoryEntry:
      type: object
      required: [toStatus, actorDisplayName, occurredAt]
      properties:
        fromStatus: { oneOf: [{ $ref: '#/components/schemas/LeaveStatus' }, { type: 'null' }] }
        toStatus: { $ref: '#/components/schemas/LeaveStatus' }
        actorDisplayName: { type: string }
        comment: { type: [string, 'null'] }
        occurredAt: { type: string, format: date-time }
    LeaveRequestPage:
      type: object
      required: [content, page, size, totalElements, totalPages]
      properties:
        content:
          type: array
          items: { $ref: '#/components/schemas/LeaveRequestSummary' }
        page: { type: integer }
        size: { type: integer }
        totalElements: { type: integer, format: int64 }
        totalPages: { type: integer }
...
    EmployeeUpdateCommand:
      type: object
      additionalProperties: false
      required: [employeeNumber, displayName, email, login, roles, active, expectedVersion]
      properties:
        employeeNumber: { type: string, minLength: 1, maxLength: 64 }
        displayName: { type: string, minLength: 1, maxLength: 160 }
        email: { type: string, format: email, maxLength: 254 }
        login: { type: string, minLength: 1, maxLength: 254 }
        managerId: { type: [string, 'null'], format: uuid }
        roles: { $ref: '#/components/schemas/Roles' }
        active: { type: boolean }
        expectedVersion: { type: integer, format: int64, minimum: 0 }
    Employee:
      type: object
      additionalProperties: false
      required: [id, employeeNumber, displayName, email, roles, active, version]
      properties:
        id: { type: string, format: uuid }
        employeeNumber: { type: string }
        displayName: { type: string }
        email: { type: string, format: email }
        managerId: { type: [string, 'null'], format: uuid }
        managerName: { type: [string, 'null'] }
        roles: { $ref: '#/components/schemas/Roles' }
        active: { type: boolean }
        version: { type: integer, format: int64 }
...
    FieldError:
      type: object
      required: [field, code, message]
      properties:
        field: { type: string }
        code: { type: string, description: Stable code such as STALE_VERSION, LEAVE_OVERLAP, INSUFFICIENT_BALANCE, or INVALID_STATUS_TRANSITION }
        message: { type: string }
    Problem:
      type: object
      required: [type, title, status, code, detail, correlationId]
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer }
        code: { type: string }
        detail: { type: string }
        correlationId: { type: string }
        fieldErrors:
          type: array
          items: { $ref: '#/components/schemas/FieldError' }